move code out of __init__.py
